requirements.txt
streamlit==1.20.0
pandas==1.5.3
psycopg2-binary==2.9.3
scikit-learn==1.2.0
